a hard days work

+++ include/createAuthFiles.php
@@ -27,64 +27,90 @@ function createAuthUserFile( $dbh ) {
 	$tMessage							= "";
 	$dir								= dirname( $CONF['AuthUserFile'] );
 	$entropy							= create_salt();
-	$tempfile							= $dir."/authtemp".$entropy;
-	
+	$tempfile							= $dir."/authtemp_".$entropy;
+		
 	if( $CONF['createUserFile'] == "YES" ) {
 		
-		if( $fileHandle	= @fopen( $tempfile, 'w' ) ) {
+		if( db_set_semaphore( 'createauthuserfile', 'sem', $dbh ) ) {
 			
-			$query						= "SELECT * " .
-										  "  FROM svnusers " .
-										  " WHERE (deleted = '0000-00-00 00:00:00') " .
-										  "   AND (locked = '0') " .
-										  "ORDER BY userid";
-			$result						= db_query( $query, $dbh );
-			
-			while( $row = db_array( $result['result'] ) ) {
+			if( $fileHandle	= @fopen( $tempfile, 'w' ) ) {
 				
-				if( ! @fwrite( $fileHandle, $row['userid'].":".$row['password']."\n" ) ) {
+				$query						= "SELECT * " .
+											  "  FROM svnusers " .
+											  " WHERE (deleted = '0000-00-00 00:00:00') " .
+											  "   AND (locked = '0') " .
+											  "ORDER BY userid";
+				$result						= db_query( $query, $dbh );
+				
+				while( $row = db_array( $result['result'] ) ) {
 					
-					$retcode 			= 1;
-					$tMessage			= _( "Can't write to AuthUser file" );
+					if( ! @fwrite( $fileHandle, $row['userid'].":".$row['password']."\n" ) ) {
+						
+						$retcode 			= 1;
+						$tMessage			= _( "Can't write to AuthUser file" );
+						db_unset_semaphore( 'createauthuserfile', 'sem', $dbh );
+					}
+					
 				}
 				
-			}
-			
-			@fclose( $fileHandle );
-			
-			if( @copy( $tempfile, $CONF['AuthUserFile'] ) ) {
+				@fclose( $fileHandle );
 				
-				if( @unlink( $tempfile ) ) {
+				if( $retcode == 0 ) {
 					
-					$tMessage				= _("Auth user file successfully created!" );
+					if( @copy( $tempfile, $CONF['AuthUserFile'] ) ) {
+						
+						if( @unlink( $tempfile ) ) {
+							
+							if( db_unset_semaphore( 'createauthuserfile', 'sem', $dbh ) ) {
+								
+								$tMessage			= _("Auth user file successfully created!" );
+								
+							} else {
+								
+								$retcode			= 1;
+								$tMessage			= _("Auth user file created but semaphore could not be released");
+								db_unset_semaphore( 'createauthuserfile', 'sem', $dbh );
+							}
+							
+						} else {
+							
+							$retcode				= 4;
+							$tMessage				= sprintf( _("Delete of %s failed!"), $tempfile );
+							db_unset_semaphore( 'createauthuserfile', 'sem', $dbh );
+						}
+						
+					} else {
+						
+						$retcode					= 3;
+						$tMessage					= sprintf( _("Copy from %s to %s failed!"), $tempfile, $CONF['AuthUserFile'] );
+						db_unset_semaphore( 'createauthuserfile', 'sem', $dbh );
+					}
 					
-				} else {
-					
-					$retcode				= 4;
-					$tMessage				= sprintf( _("Delete of %s failed!"), $tempfile );
 				}
 				
 			} else {
 				
-				$retcode					= 3;
-				$tMessage					= sprintf( _("Copy from %s to %s failed!"), $tempfile, $CONF['AuthUserFile'] );
+				$retcode						= 2;
+				$tMessage						= sprintf( _( "Cannot open file %s for writing!" ), $tempfile );
+				db_unset_semaphore( 'createauthuserfile', 'sem', $dbh );
 			}
 			
 		} else {
-			
-			$retcode						= 2;
-			$tMessage						= sprintf( _( "Cannot open file %s for writing!" ), $tempfile );
+		
+			$error								= 1;
+			$tMessage							= _("Can't set semaphore, another process is writing Auth User File, try again later");
 		}
 		
 	} else {
 		
-		$retcode							= 0;
-		$tMessage							= _("Create of auth user file not configured!" );
+		$retcode								= 0;
+		$tMessage								= _("Create of auth user file not configured!" );
 	}
+		
 	
-	$ret									= array();
-	$ret['error']							= $retcode;
-	$ret['errormsg']						= $tMessage;
+	$ret										= array();
+	$ret['error']								= $retcode;
+	$ret['errormsg']							= $tMessage;
 	
 	return $ret;
 }
@@ -101,184 +127,213 @@ function createAccessFile( $dbh ) {
 	
 	if( $CONF['createAccessFile'] == "YES" ) {
 		
-		$dir							= dirname( $CONF['SVNAccessFile'] );
-		$entropy						= create_salt();
-		$tempfile						= $dir."/accesstemp".$entropy;
-	
-		if( $fileHandle = @fopen ( $tempfile, 'w' ) ) {
+		if( db_set_semaphore( 'createaccessfile', 'sem', $dbh ) ) {
+			
+			$dir							= dirname( $CONF['SVNAccessFile'] );
+			$entropy						= create_salt();
+			$tempfile						= $dir."/accesstemp_".$entropy;
 		
-			if( ! @fwrite( $fileHandle, "[groups]\n" ) ) {
-				
-				$retcode					= 1;
-				$tMessage					= sprintf( _("Cannot write to %s"), $tempfile );
-			} 
+			if( $fileHandle = @fopen ( $tempfile, 'w' ) ) {
 			
-			if( $retcode == 0 ) {
-			
-				# write groups to file
-				$query							= "  SELECT svngroups.groupname, svnusers.userid " .
-												  "    FROM svngroups, svnusers, svn_users_groups " .
-												  "   WHERE (svngroups.deleted = '0000-00-00 00:00:00') " .
-												  "     AND (svn_users_groups.user_id = svnusers.id) " .
-												  "     AND (svn_users_groups.group_id = svngroups.id) " .
-												  "     AND (svnusers.deleted = '0000-00-00 00:00:00') " .
-												  "     AND (svn_users_groups.deleted = '0000-00-00 00:00:00') " .
-												  "ORDER BY svngroups.groupname ASC";
-				$result							= db_query( $query, $dbh );
-				$oldgroup						= "";
-				$users							= "";
-				
-				while( ($row = db_array( $result['result'] )) and ($retcode == 0) ) {
+				if( ! @fwrite( $fileHandle, "[groups]\n" ) ) {
 					
-					if( $oldgroup != $row['groupname'] ) {
+					$retcode					= 1;
+					$tMessage					= sprintf( _("Cannot write to %s"), $tempfile );
+					db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
+				} 
+				
+				if( $retcode == 0 ) {
+				
+					# write groups to file
+					$query							= "  SELECT svngroups.groupname, svnusers.userid " .
+													  "    FROM svngroups, svnusers, svn_users_groups " .
+													  "   WHERE (svngroups.deleted = '0000-00-00 00:00:00') " .
+													  "     AND (svn_users_groups.user_id = svnusers.id) " .
+													  "     AND (svn_users_groups.group_id = svngroups.id) " .
+													  "     AND (svnusers.deleted = '0000-00-00 00:00:00') " .
+													  "     AND (svn_users_groups.deleted = '0000-00-00 00:00:00') " .
+													  "ORDER BY svngroups.groupname ASC";
+					$result							= db_query( $query, $dbh );
+					$oldgroup						= "";
+					$users							= "";
+					
+					while( ($row = db_array( $result['result'] )) and ($retcode == 0) ) {
 						
-						if( $users != "" ) {
+						if( $oldgroup != $row['groupname'] ) {
 							
-							if( ! @fwrite( $fileHandle, $oldgroup." = ".$users."\n" ) )  {
+							if( $users != "" ) {
 								
-								$retcode		= 1;
-								$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
+								if( ! @fwrite( $fileHandle, $oldgroup." = ".$users."\n" ) )  {
+									
+									$retcode		= 1;
+									$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
+									db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
+								}
+								
+							} 
+							
+							$users					= $row['userid'];
+							$oldgroup				= $row['groupname'];
+							
+						} else {
+							
+							if( $users == "" ) {
+								
+								$users				= $row['userid'];
+								
+							} else {
+								
+								$users				= $users.", ".$row['userid'];
+								
+							}
+							
+						}
+						
+					}
+					
+					if( $users != "" ) {
+						
+						fwrite( $fileHandle, $oldgroup." = ".$users."\n" );
+						
+					}
+				}
+				
+				
+				if( $retcode == 0 ) {
+				
+					# write access rights to file
+					$query							= "  SELECT svnmodule, modulepath, reponame, path, user_id, group_id, access_right " .
+													  "    FROM svn_access_rights, svnprojects, svnrepos " .
+													  "   WHERE (svn_access_rights.deleted = '0000-00-00 00:00:00') " .
+													  "     AND (svn_access_rights.valid_from <= '$curdate') " .
+													  "     AND (svn_access_rights.valid_until >= '$curdate') " .
+													  "     AND (svn_access_rights.project_id = svnprojects.id) " .
+													  "     AND (svnprojects.repo_id = svnrepos.id) " .
+													  "ORDER BY svn_access_rights.path ASC";
+					$result							= db_query( $query, $dbh );
+					
+					while( ($row = db_array( $result['result'] )) and ($retcode == 0) ) {
+						
+						if( $row['access_right'] == "none" ) {
+							
+							$right					= "";
+							
+						} elseif( $row['access_right'] == "read" ) {
+							
+							$right					= "r";
+							
+						} elseif( $row['access_right'] == "write" ) {
+							
+							$right					= "rw";
+							
+						} else {
+							
+							$right					= "";
+							
+						}
+						
+						if( $row['path'] != $oldpath ) {
+							
+							$oldpath				= $row['path'];
+							if( ! @fwrite( $fileHandle, "\n[".$row['reponame'].":".$row['path']."]\n" ) ) {
+								
+								$retcode			= 4;
+								$tMessage			= sprintf( _("Cannot write to %s"), $tempfile );
+								db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
 							}
 							
 						} 
 						
-						$users					= $row['userid'];
-						$oldgroup				= $row['groupname'];
-						
-					} else {
-						
-						if( $users == "" ) {
+						if( $row['user_id'] != "0" ) {
 							
-							$users				= $row['userid'];
+							$query					= "SELECT * " .
+													  "  FROM svnusers " .
+													  " WHERE (id = ".$row['user_id'].")";
+							$resultusr				= db_query( $query, $dbh );
+							
+							if( $resultusr['rows'] == 1 ) {
+								
+								$rowusr				= db_array( $resultusr['result'] );
+								if( ! @fwrite( $fileHandle, $rowusr['userid']." = ".$right."\n" ) ) {
+									
+									$retcode		= 5;
+									$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
+									db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
+								}
+								
+							}
+						}
+						
+						if( $row['group_id'] != "0" ) {
+						
+							$query					= "  SELECT * " .
+													  "    FROM svngroups " .
+													  "   WHERE (id = ".$row['group_id'].")";
+							$resultgrp				= db_query( $query, $dbh );
+							
+							if( $resultgrp['rows'] == 1 ) {
+								
+								$rowgrp				= db_array( $resultgrp['result'] );
+								if( ! @fwrite( $fileHandle, "@".$rowgrp['groupname']." = ".$right."\n" ) ) {
+									
+									$retcode		= 6;
+									$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
+									db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
+								}
+								
+							} 	
+						}
+					}
+					
+				}
+				
+				@fclose( $fileHandle );
+				
+				if( @copy( $tempfile, $CONF['SVNAccessFile'] ) ) {
+					
+					if( @unlink( $tempfile ) ) {
+						
+						if( db_unset_semaphore( 'createaccessfile', 'sem', $dbh ) ) {
+						
+							$tMessage				= _( "Access file successfully created!" );
 							
 						} else {
 							
-							$users				= $users.", ".$row['userid'];
+							$error					= 1;
+							$tMessage				= _("Access file successfully created but semaphore could nor be released");
+							db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
 							
-						}
-						
-					}
-					
-				}
-				
-				if( $users != "" ) {
-					
-					fwrite( $fileHandle, $oldgroup." = ".$users."\n" );
-					
-				}
-			}
-			
-			
-			if( $retcode == 0 ) {
-			
-				# write access rights to file
-				$query							= "  SELECT svnmodule, modulepath, reponame, path, user_id, group_id, access_right " .
-												  "    FROM svn_access_rights, svnprojects, svnrepos " .
-												  "   WHERE (svn_access_rights.deleted = '0000-00-00 00:00:00') " .
-												  "     AND (svn_access_rights.valid_from <= '$curdate') " .
-												  "     AND (svn_access_rights.valid_until >= '$curdate') " .
-												  "     AND (svn_access_rights.project_id = svnprojects.id) " .
-												  "     AND (svnprojects.repo_id = svnrepos.id) " .
-												  "ORDER BY svn_access_rights.path ASC";
-				$result							= db_query( $query, $dbh );
-				
-				while( ($row = db_array( $result['result'] )) and ($retcode == 0) ) {
-					
-					if( $row['access_right'] == "none" ) {
-						
-						$right					= "";
-						
-					} elseif( $row['access_right'] == "read" ) {
-						
-						$right					= "r";
-						
-					} elseif( $row['access_right'] == "write" ) {
-						
-						$right					= "rw";
+						}	
 						
 					} else {
 						
-						$right					= "";
-						
+						$retcode				= 4;
+						$tMessage				= sprintf( _("Delete of %s failed!"), $tempfile );
+						db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
 					}
 					
-					if( $row['path'] != $oldpath ) {
-						
-						$oldpath				= $row['path'];
-						if( ! @fwrite( $fileHandle, "\n[".$row['reponame'].":".$row['path']."]\n" ) ) {
-							
-							$retcode			= 4;
-							$tMessage			= sprintf( _("Cannot write to %s"), $tempfile );
-						}
-						
-					} 
-					
-					if( $row['user_id'] != "0" ) {
-						
-						$query					= "SELECT * " .
-												  "  FROM svnusers " .
-												  " WHERE (id = ".$row['user_id'].")";
-						$resultusr				= db_query( $query, $dbh );
-						
-						if( $resultusr['rows'] == 1 ) {
-							
-							$rowusr				= db_array( $resultusr['result'] );
-							if( ! @fwrite( $fileHandle, $rowusr['userid']." = ".$right."\n" ) ) {
-								
-								$retcode		= 5;
-								$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
-							}
-							
-						}
-					}
-					
-					if( $row['group_id'] != "0" ) {
-					
-						$query					= "  SELECT * " .
-												  "    FROM svngroups " .
-												  "   WHERE (id = ".$row['group_id'].")";
-						$resultgrp				= db_query( $query, $dbh );
-						
-						if( $resultgrp['rows'] == 1 ) {
-							
-							$rowgrp				= db_array( $resultgrp['result'] );
-							if( ! @fwrite( $fileHandle, "@".$rowgrp['groupname']." = ".$right."\n" ) ) {
-								
-								$retcode		= 6;
-								$tMessage		= sprintf( _("Cannot write to %s"), $tempfile );
-							}
-							
-						} 	
-					}
-				}
-				
-			}
-			
-			@fclose( $fileHandle );
-			
-			if( @copy( $tempfile, $CONF['SVNAccessFile'] ) ) {
-				
-				if( @unlink( $tempfile ) ) {
-					
-					$tMessage				= _( "Access file successfully created!" );	
 				} else {
 					
-					$retcode				= 4;
-					$tMessage				= sprintf( _("Delete of %s failed!"), $tempfile );
+					$retcode					= 3;
+					$tMessage					= sprintf( _("Copy from %s to %s failed!"), $tempfile, $CONF['SVNAccessFile'] );
+					db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
 				}
-				
+			
 			} else {
 				
-				$retcode					= 3;
-				$tMessage					= sprintf( _("Copy from %s to %s failed!"), $tempfile, $CONF['SVNAccessFile'] );
+				$retcode						= 1;
+				$tMessage						= sprintf( _("Cannot open %s for wrtiting"), $tempfile );
+				db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
+				
 			}
-		
+			
 		} else {
 			
-			$retcode						= 1;
-			$tMessage						= sprintf( _("Cannot open %s for wrtiting"), $tempfile );
-			
+			$error								= 1;
+			$tMessage							= _("Can't set semaphore, another process is writing access file, try again later");
+			db_unset_semaphore( 'createaccessfile', 'sem', $dbh );
 		}
+	
 	} else {
 		
 		$retcode							= 0;
